feat: профиль

+++ calendar/src/main/java/com/coursework/calendar/service/InvitationService.java
@@ -6,6 +6,7 @@ import java.util.UUID;
 import org.springframework.stereotype.Service;
 
 import com.coursework.calendar.entities.invitation.Invitation;
+import com.coursework.calendar.entities.invitation.InvitationStatus;
 import com.coursework.calendar.repository.InvitationRepository;
 
 @Service
@@ -40,6 +41,10 @@ public class InvitationService {
     }
 
     public void deleteInvitation(UUID id) {
+        if (!invitationRepository.existsById(id)) {
+            throw new RuntimeException("Invitation not found");
+        }
+
         invitationRepository.deleteById(id);
     }
 }
